🧹 Code Health: Extract duplicated validation helpers
- Moves `validateSlug`, `validateName`, `validateDescription`, and `SLUG_RE` from `routes/orgs.ts` and `routes/collections.ts` into a new `utils/validation.ts` file.
- Updates routing files to import these helpers.
